ground_for_the_house
As the input data you will get the multiline string consists of '0' & '#'. where '0' means the empty piece of the ground and the '#' is the piece of your house. Your task is to count the minimal area of the rectangle ground which is enough for the building.
Input: The plan of the house.
Output: The total area of the rectangle piece of the ground.